chore: separate sp.oidc.client.cache in separate file
